requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
idna==3.10
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
django-redis==5.4.0

